Applying changes from user prompt: The assistant.go file has gotten a bit unwieldly. Could you move the functions related to performing git operations into a separate file, git_operations.go?
